requirements.txt
# Основные зависимости
fastapi==0.109.2
uvicorn==0.27.0

# База данных и ORM
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Аутентификация и безопасность
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19

# Валидация данных
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator==2.1.1

# Дополнительные утилиты
python-dotenv==1.0.1
loguru==0.7.2
typing-extensions==4.10.0

# Для разработки
pytest==8.0.2
pytest-asyncio==0.23.5
httpx==0.26.0
sqlalchemy2-stubs==0.0.2a35